Known bugs in this version:
Simple4 is writing/reading pages to the disk, even though they were only read. Causes page ins/outs to be too high, as well as switches.

simple5 is off by ~100 switches, but everything else is literally identical. Not convicned this is a problem we need to worry about

Everything else works, not mentioning the constan bus error, but we can also ingnore those